Allow tutorials to be sorted by the json config
Tutorial metadata now has an autoindex field, which is based on the
order in which the tutorials appear in the json config file.

Templates can choose to sort based on this if they wish.

This adds the "sortTutorialsByConfigOrder" option to the default
template to sort based on this new field.

Issue #1028